+++ tests/src/test/java/org/sonarqube/tests/Category6Suite.java
@@ -26,17 +26,11 @@ import org.junit.ClassRule;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.sonarqube.tests.authorisation.PermissionTemplateTest;
-import org.sonarqube.tests.issue.IssueNotificationsTest;
 import org.sonarqube.tests.ce.ReportFailureNotificationTest;
+import org.sonarqube.tests.issue.IssueNotificationsTest;
 import org.sonarqube.tests.issue.IssueTagsTest;
 import org.sonarqube.tests.issue.OrganizationIssueAssignTest;
 import org.sonarqube.tests.issue.OrganizationIssuesPageTest;
-import org.sonarqube.tests.organization.BillingTest;
-import org.sonarqube.tests.organization.OrganizationMembershipTest;
-import org.sonarqube.tests.organization.OrganizationMembershipUiTest;
-import org.sonarqube.tests.organization.OrganizationTest;
-import org.sonarqube.tests.organization.PersonalOrganizationTest;
-import org.sonarqube.tests.organization.RootUserOnOrganizationTest;
 import org.sonarqube.tests.projectAdministration.ProjectDeleteTest;
 import org.sonarqube.tests.projectAdministration.ProjectKeyUpdateTest;
 import org.sonarqube.tests.projectAdministration.ProjectProvisioningTest;
@@ -50,7 +44,6 @@ import org.sonarqube.tests.qualityProfile.OrganizationQualityProfilesUiTest;
 import org.sonarqube.tests.qualityProfile.QualityProfilesEditTest;
 import org.sonarqube.tests.qualityProfile.QualityProfilesWsTest;
 import org.sonarqube.tests.rule.RulesWsTest;
-import org.sonarqube.tests.ui.OrganizationUiExtensionsTest;
 import org.sonarqube.tests.user.OrganizationIdentityProviderTest;
 
 import static util.ItUtils.pluginArtifact;
@@ -64,19 +57,12 @@ import static util.ItUtils.xooPlugin;
   OrganizationIdentityProviderTest.class,
   OrganizationIssueAssignTest.class,
   OrganizationIssuesPageTest.class,
-  OrganizationMembershipTest.class,
-  OrganizationMembershipUiTest.class,
   OrganizationQualityGateUiTest.class,
   OrganizationQualityProfilesUiTest.class,
-  OrganizationTest.class,
-  RootUserOnOrganizationTest.class,
-  OrganizationUiExtensionsTest.class,
-  PersonalOrganizationTest.class,
   BuiltInQualityProfilesTest.class,
   QualityProfilesEditTest.class,
   QualityProfilesWsTest.class,
   CustomQualityProfilesTest.class,
-  BillingTest.class,
   IssueTagsTest.class,
   LeakProjectsPageTest.class,
   SearchProjectsTest.class,
@@ -104,7 +90,6 @@ public class Category6Suite {
 
     .addPlugin(xooPlugin())
     .addPlugin(pluginArtifact("base-auth-plugin"))
-    .addPlugin(pluginArtifact("fake-billing-plugin"))
     .addPlugin(pluginArtifact("ui-extensions-plugin"))
 
     // reduce memory for Elasticsearch to 128M
